[IOTMID-946] Fixes config for empty databases
Within the generated container image, the service is managed
by supervisord and run through uwsgi. Since uwsgi does not call
the module as a binary, the conditional __main__ bit of app.py
is never run, thus leaving the database empty on the "official"
environment.

Since getConfValue() would not check whether the config collection
was empty, it would fail on those pristine environments, leading
to the reported failure to create the admin (initial) user.

This change adds proper handling for the failure of find_one(),
by forcibly initializing the configuration collection if it
doesn't exist.

Change-Id: I2f0b34ef1ea7b89e38d76ceeb5892543ec3e330d