Adds deserializer for timestamps, fixes sanity checks
This commit is two-fold: firstly it provides a cleaner alternative
for bson.json_util, which relied on the availability of pymongo,
which has been dropped as a dependency.

Secondly, this fixes a couple of typos on permissions checks, where
the URI path would end up being validated as if it were the permission
human-readable label.